+++ pkg/metricsservice/client.go
@@ -65,24 +65,19 @@ func NewGrpcClient(url, certDir string) (*GrpcClient, error) {
 	return &GrpcClient{client: api.NewMetricsServiceClient(conn), connection: conn}, nil
 }
 
-func (c *GrpcClient) GetMetrics(ctx context.Context, scaledObjectName, scaledObjectNamespace, metricName string) (*external_metrics.ExternalMetricValueList, *api.PromMetricsMsg, error) {
-	// nosemgrep: trailofbits.go.invalid-usage-of-modified-variable.invalid-usage-of-modified-variable
-	response, err := c.client.GetMetrics(ctx, &api.ScaledObjectRef{Name: scaledObjectName, Namespace: scaledObjectNamespace, MetricName: metricName})
+func (c *GrpcClient) GetMetrics(ctx context.Context, scaledObjectName, scaledObjectNamespace, metricName string) (*external_metrics.ExternalMetricValueList, error) {
+	v1beta1ExtMetrics, err := c.client.GetMetrics(ctx, &api.ScaledObjectRef{Name: scaledObjectName, Namespace: scaledObjectNamespace, MetricName: metricName})
 	if err != nil {
-		// in certain cases we would like to get Prometheus metrics even if there's an error
-		// so we can expose information about the error in the client
-		return nil, response.GetPromMetrics(), err
+		return nil, err
 	}
 
 	extMetrics := &external_metrics.ExternalMetricValueList{}
-	err = v1beta1.Convert_v1beta1_ExternalMetricValueList_To_external_metrics_ExternalMetricValueList(response.GetMetrics(), extMetrics, nil)
+	err = v1beta1.Convert_v1beta1_ExternalMetricValueList_To_external_metrics_ExternalMetricValueList(v1beta1ExtMetrics, extMetrics, nil)
 	if err != nil {
-		// in certain cases we would like to get Prometheus metrics even if there's an error
-		// so we can expose information about the error in the client
-		return nil, response.GetPromMetrics(), fmt.Errorf("error when converting metric values %w", err)
+		return nil, fmt.Errorf("error when converting metric values %w", err)
 	}
 
-	return extMetrics, response.GetPromMetrics(), nil
+	return extMetrics, nil
 }
 
 // WaitForConnectionReady waits for gRPC connection to be ready